feat: add query types for TrustUpdate & SeedUpdate download
